LayoutTests/platform/blackberry/accessibility/aria-readonly-expected.txt

     
This tests that the aria-readonly attribute works. The first and third text fields should not be writable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is false
FAIL succeeded should be true. Was false.
PASS succeeded is false
FAIL succeeded should be true. Was false.
PASS succeeded is false
FAIL succeeded should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
